Updated tox.ini to run test_SandhiKosh, removed uohd tests from test_SanskritLexicalAnalyzer.py, since they're in SandhiKosh
