+++ tox.ini
@@ -11,7 +11,7 @@ deps =
     pytest
     google_compute_engine
 commands =
-    pytest --test-count 1000 tests/test_DhatuWrapper.py tests/test_sandhi.py tests/test_SanskritLexicalAnalyer.py
+    pytest --test-count 1000 tests/test_DhatuWrapper.py tests/test_sandhi.py tests/test_SanskritLexicalAnalyzer.py tests/test_SandhiKosh.py
 
 [testenv:lint]
 basepython = python
